@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -37,9 +39,9 @@
 		"bluebird"
 	],
 	"devDependencies": {
-		"ava": "^1.4.1",
-		"sinon": "^7.3.1",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"ava": "^3.15.0",
+		"sinon": "^10.0.0",
+		"tsd": "^0.14.0",
+		"xo": "^0.38.2"
 	}
 }
