@@ -1,12 +1,11 @@
-declare namespace pPipe {
-	type UnaryFunction<ValueType, ReturnType> = (
-		value: ValueType
-	) => ReturnType | PromiseLike<ReturnType>;
+/* eslint-disable import/export */
+export type UnaryFunction<ValueType, ReturnType> = (
+	value: ValueType
+) => ReturnType | PromiseLike<ReturnType>;
 
-	type Pipeline<ValueType, ReturnType> = (
-		value?: ValueType
-	) => Promise<ReturnType>;
-}
+export type Pipeline<ValueType, ReturnType> = (
+	value?: ValueType
+) => Promise<ReturnType>;
 
 /**
 Compose promise-returning & async functions into a reusable pipeline.
@@ -16,44 +15,42 @@ Compose promise-returning & async functions into a reusable pipeline.
 
 @example
 ```
-import pPipe = require('p-pipe');
+import pPipe from 'p-pipe';
 
 const addUnicorn = async string => `${string} Unicorn`;
 const addRainbow = async string => `${string} Rainbow`;
 
 const pipeline = pPipe(addUnicorn, addRainbow);
 
-(async () => {
-	console.log(await pipeline('❤️'));
-	//=> '❤️ Unicorn Rainbow'
-})();
+console.log(await pipeline('❤️'));
+//=> '❤️ Unicorn Rainbow'
 ```
 */
-declare function pPipe<ValueType, ReturnType>(
-	f1: pPipe.UnaryFunction<ValueType, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<ValueType, ResultValue1, ReturnType>(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<ValueType, ResultValue1, ResultValue2, ReturnType>(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+export default function pPipe<ValueType, ReturnType>(
+	f1: UnaryFunction<ValueType, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<ValueType, ResultValue1, ReturnType>(
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<ValueType, ResultValue1, ResultValue2, ReturnType>(
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
 	ResultValue3,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
@@ -61,13 +58,13 @@ declare function pPipe<
 	ResultValue4,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ResultValue4>,
-	f5: pPipe.UnaryFunction<ResultValue4, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ResultValue4>,
+	f5: UnaryFunction<ResultValue4, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
@@ -76,14 +73,14 @@ declare function pPipe<
 	ResultValue5,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ResultValue4>,
-	f5: pPipe.UnaryFunction<ResultValue4, ResultValue5>,
-	f6: pPipe.UnaryFunction<ResultValue5, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ResultValue4>,
+	f5: UnaryFunction<ResultValue4, ResultValue5>,
+	f6: UnaryFunction<ResultValue5, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
@@ -93,15 +90,15 @@ declare function pPipe<
 	ResultValue6,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ResultValue4>,
-	f5: pPipe.UnaryFunction<ResultValue4, ResultValue5>,
-	f6: pPipe.UnaryFunction<ResultValue5, ResultValue6>,
-	f7: pPipe.UnaryFunction<ResultValue6, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ResultValue4>,
+	f5: UnaryFunction<ResultValue4, ResultValue5>,
+	f6: UnaryFunction<ResultValue5, ResultValue6>,
+	f7: UnaryFunction<ResultValue6, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
@@ -112,16 +109,16 @@ declare function pPipe<
 	ResultValue7,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ResultValue4>,
-	f5: pPipe.UnaryFunction<ResultValue4, ResultValue5>,
-	f6: pPipe.UnaryFunction<ResultValue5, ResultValue6>,
-	f7: pPipe.UnaryFunction<ResultValue6, ResultValue7>,
-	f8: pPipe.UnaryFunction<ResultValue7, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
-declare function pPipe<
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ResultValue4>,
+	f5: UnaryFunction<ResultValue4, ResultValue5>,
+	f6: UnaryFunction<ResultValue5, ResultValue6>,
+	f7: UnaryFunction<ResultValue6, ResultValue7>,
+	f8: UnaryFunction<ResultValue7, ReturnType>
+): Pipeline<ValueType, ReturnType>;
+export default function pPipe<
 	ValueType,
 	ResultValue1,
 	ResultValue2,
@@ -133,20 +130,18 @@ declare function pPipe<
 	ResultValue8,
 	ReturnType
 >(
-	f1: pPipe.UnaryFunction<ValueType, ResultValue1>,
-	f2: pPipe.UnaryFunction<ResultValue1, ResultValue2>,
-	f3: pPipe.UnaryFunction<ResultValue2, ResultValue3>,
-	f4: pPipe.UnaryFunction<ResultValue3, ResultValue4>,
-	f5: pPipe.UnaryFunction<ResultValue4, ResultValue5>,
-	f6: pPipe.UnaryFunction<ResultValue5, ResultValue6>,
-	f7: pPipe.UnaryFunction<ResultValue6, ResultValue7>,
-	f8: pPipe.UnaryFunction<ResultValue7, ResultValue8>,
-	f9: pPipe.UnaryFunction<ResultValue8, ReturnType>
-): pPipe.Pipeline<ValueType, ReturnType>;
+	f1: UnaryFunction<ValueType, ResultValue1>,
+	f2: UnaryFunction<ResultValue1, ResultValue2>,
+	f3: UnaryFunction<ResultValue2, ResultValue3>,
+	f4: UnaryFunction<ResultValue3, ResultValue4>,
+	f5: UnaryFunction<ResultValue4, ResultValue5>,
+	f6: UnaryFunction<ResultValue5, ResultValue6>,
+	f7: UnaryFunction<ResultValue6, ResultValue7>,
+	f8: UnaryFunction<ResultValue7, ResultValue8>,
+	f9: UnaryFunction<ResultValue8, ReturnType>
+): Pipeline<ValueType, ReturnType>;
 
 // Fallbacks if more than 9 functions are passed as input (not type-safe).
-declare function pPipe(
-	...functions: (pPipe.UnaryFunction<any, unknown>)[]
-): pPipe.Pipeline<unknown, unknown>;
-
-export = pPipe;
+export default function pPipe(
+	...functions: Array<UnaryFunction<any, unknown>>
+): Pipeline<unknown, unknown>;
