@@ -1,6 +1,4 @@
-'use strict';
-
-module.exports = (...functions) => {
+export default function pPipe(...functions) {
 	if (functions.length === 0) {
 		throw new Error('Expected at least one argument');
 	}
@@ -8,10 +6,10 @@ module.exports = (...functions) => {
 	return async input => {
 		let currentValue = input;
 
-		for (const fn of functions) {
-			currentValue = await fn(currentValue); // eslint-disable-line no-await-in-loop
+		for (const function_ of functions) {
+			currentValue = await function_(currentValue); // eslint-disable-line no-await-in-loop
 		}
 
 		return currentValue;
 	};
-};
+}
